// microsoft-365/src/planner/types.ts
export interface PlannerChecklistItem {
  "@odata.type"?: string;
  title?: string;
  isChecked?: boolean;
  orderHint?: string;
  lastModifiedDateTime?: string;
  lastModifiedBy?: { user?: { id?: string } };
}

export type PlannerChecklist = Record<string, PlannerChecklistItem>;

export interface PlannerExternalReference {
  "@odata.type"?: string;
  alias?: string;
  type?: string;
  previewPriority?: string;
  lastModifiedDateTime?: string;
  lastModifiedBy?: { user?: { id?: string } };
}

export type PlannerReferences = Record<string, PlannerExternalReference>;